Turns out the driver itself knows what it supports...
Due to digging into the 4.3.1 code, and making it work for
the 4.9.0 kernel I found that the PCI IDs are in the pci.c
file. It's actually shown there what is supported and what
the driver will load on.
In the newer driver the IDs have codenames, instead of their
device name.

3.2.15:
struct pci_device_id iodrive_ids[] = {
    {PCI_DEVICE (PCI_VENDOR_ID_FUSION, IODRIVE_PCI_DEVICE)       },
    {PCI_DEVICE (PCI_VENDOR_ID_FUSION, IODRIVE1_PCI_DEVICE)      },
    {PCI_DEVICE (PCI_VENDOR_ID_FUSION, IODRIVE3_OLD_PCI_DEVICE)  },
    {PCI_DEVICE (PCI_VENDOR_ID_FUSION, IODRIVE3_BOOT_PCI_DEVICE) },
    {PCI_DEVICE (PCI_VENDOR_ID_FUSION, IODRIVE3_PCI_DEVICE)      },
    {PCI_DEVICE (PCI_VENDOR_ID_FUSION, IOXTREME_PCI_DEVICE)      },
    {PCI_DEVICE (PCI_VENDOR_ID_FUSION, IOXTREMEPRO_PCI_DEVICE)   },
    {PCI_DEVICE (PCI_VENDOR_ID_FUSION, IOXTREME2_PCI_DEVICE)     },
    {PCI_DEVICE (PCI_VENDOR_ID_FUSION, IODRIVE4_PCI_DEVICE)      },
    {0,}
};

VS

4.3.1:
struct pci_device_id iodrive_ids[] = {
    {PCI_DEVICE (PCI_VENDOR_ID_FUSION, IOSCALE3_ORANGE_PCI_DEVICE)      },
    {PCI_DEVICE (PCI_VENDOR_ID_FUSION, IOSCALE3_TANGERINE_PCI_DEVICE)   },
    {PCI_DEVICE (PCI_VENDOR_ID_FUSION, IOSCALE3_PEACH_PLUM_PCI_DEVICE)  },
    {PCI_DEVICE (PCI_VENDOR_ID_FUSION, IOSCALE3_APRICOT_PCI_DEVICE)     },
    {0,}
};

The original devices are all still in the pci.h file located in
include/fio/port.a

on to testing...